requirements.txt.txt
Django==4.2.5
xmltodict==0.13.0
pycryptodome==3.20.0
asgiref==3.7.2
sqlparse==0.3.1
typing_extensions==4.9.0
tabula-py==2.9.0